Moved to new stateful code, need to fix callbacks
